\name{BCEs0-package}
\alias{BCEs0-package}
\alias{BCEs0}
\docType{package}
\title{
BCEs0. A R package to implement Bayesian models for cost-effectiveness analysis 
in the presence of structural zero costs
}
\description{
Implements a full Bayesian cost-effectiveness analysis in the case where the cost
variable is characterised by structural zeros. The package implements the Gamma, 
log-Normal and Normal models for the cost variable and the Gamma, Beta, Bernoulli 
and Normal models for the measure of clinical effectiveness. 
}
\details{
\tabular{ll}{
Package: \tab BCEs0\cr
Type: \tab Package\cr
Version: \tab 1.1\cr
Date: \tab 2013-11-15\cr
License: \tab GPL (>=2)\cr
LazyLoad: \tab yes\cr
}
}
\author{
Gianluca Baio

Maintainer: Gianluca Baio <gianluca@stats.ucl.ac.uk>
}
\references{
Baio G. (2013). Bayesian models for cost-effectiveness analysis in the presence 
of structural zero costs. http://arxiv.org/pdf/1307.5243v1.pdf
}
\examples{
## To be added here
}
\keyword{Bayesian models}
\keyword{Cost-effectiveness analysis}
\keyword{Hurdle models}
\keyword{Structural zero costs}

